ad-web-manager-1/requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
ldap3
gunicorn
python-dotenv